Don't output Container.parameters in API resulsts
The parameters store arbitrary objects for which there may no exists a
serializer.
In the case of the kube backend it also contains the full Pod object,
maybe it is not desired to output this in the API.

Bug can be reproduced by adding proxy.container-log-path=./container-logs to
the app config.